Generics
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@14497 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-kernel/src/com/liferay/portal/kernel/util/Validator.java
@@ -91,7 +91,7 @@ public class Validator {
 	}
 
 	public static boolean isDigit(char c) {
-		int x = (int)c;
+		int x = c;
 
 		if ((x >= 48) && (x <= 57)) {
 			return true;
